ch4_backend/requirements.txt
django 
django-cors-headers
django-rest-auth
djangorestframework
gunicorn
whitenoise